// back-end/src/coupon/coupon.interface.ts
export class UserCoupon {
  USR_CPN_SQ: number;
  CPN_GET_TP: number;
  USR_SQ: number;
  JOIN_USR_SQ: number;
  JOIN_USR_NM: string;
  USE_CPN_SQ: number;
  USE_CPN_DT: string;
  PRF_SQ: number;
  PRF_NM: string;
  TEST_SQ: number;
  CRT_DT: string;
  MOD_DT: string;
  DEL_DT: string;
}

export class UseCoupon {
  USE_CPN_SQ: number;
  USR_CPN_SQ: number;
  TEST_SQ: number;
  CRT_DT: string;
  MOD_DT: string;
  DEL_DT: string;
}